requirements.txt
Django==5.0.3
psycopg==3.1.18
djangorestframework==3.15.0
redis==5.0.3
celery==5.3.6
gunicorn==21.2.0
python-dotenv==1.0.1
httpx==0.27.0
pytz==2024.1
pillow==10.2.0
django-cors-headers==4.3.1
whitenoise==6.6.0
dj-database-url==2.1.0
sentry-sdk==1.42.0
PyJWT==2.8.0
aiosmtpd==1.4.5
pytest-playwright==0.4.4
django-vite==3.0.3
django-cors-headers==4.3.1
